requirements.txt
pyxel
pymunk
pysdl2
pillow
